added more logic to save positions of pre-existing and new stickies
